\name{RMongo-class}
\alias{RMongo-class}

\title{Class RMongo}

\description{
Contains the mongodb connection information. See mongoDbConnect for creating a RMongo object.
}
\section{Methods}{
  \describe{
    \item{dbGetQuery}{}
    \item{dbInsertDocument}{}
    \item{dbDisconnect}{}
  }
}

\seealso{
  \code{\link{mongoDbConnect}}
}

